Added option to check protocol availability
This is a first draft(not working yet) to implement a check for LTHOR protocol.
It's needed by our testing software. It should be able to first check if device
is in flashable mode and then perform flashing by running lthor <image.tar>.

To implement this feature correctly I need your help. Is there any way
in THOR protocol to inform the device that it should again wait for
handshake? If I just perform handshake and get information from the
device it will not be possible to flash it after that, because it will
be in active mode, waiting for the commands, not for the handshake.

Is it possible to contact you by email or this is the only communication
channel I can use?

Thank you,
Ed

Change-Id: Ic4cfa954885408343b5aad20fa805b2788c49ee7
Signed-off-by: Ed Bartosh <eduard.bartosh@intel.com>